Update pattern to use Istio 0.8
This commit modifies the configuration files to reflect new changes in
Istio 0.8. Istio Gateway is now used instead of a Kubernetes Ingress
resource to access the service mesh.

This commit also modifies the circuit breaker configuration. It is now
DestinationRule instead of DestinationPolicy. This should make Step 3
and Step 4 work.

Changes reflect the use of the new Route Rules v1alpha3,
https://istio.io/docs/reference/config/istio.networking.v1alpha3/.

Signed-off-by: AnthonyAmanse <ghieamanse@gmail.com>