Fix Istio 0.8 rules in step 5
This commit fixes the Istio timeouts and fault injection configuration
files that will be used in step 5. Previously, it was still using Route
rules v1alpha1. This commit uses Route rules v1alpha3 for Istio 0.8.

This commit also modifies the README to reflect the changes made in
those configuration files. This commit also uses the architecture README
in the IBM Developer site.

Signed-off-by: AnthonyAmanse <ghieamanse@gmail.com>