Fix circuit-breaker configuration
This commit fixes the circuit-breaker-db.yaml configuration.
`http:` is not needed for Istio 1.0. This commit also updates the README
reflecting this change.

This commit also fixes the command on creating the cloudant credentials.
Wraps the command substitution in "$(<command>)" instead of backitcks
"`<command>`". Braces are more preferred. Also adds in "LC_CTYPE=C" for
compatibility on macOS. This just changes the locale so "tr" can operate
on bytes.

Signed-off-by: AnthonyAmanse <ghieamanse@gmail.com>