Fix Step 5's timeout configuration
This commit fixes the yaml files for getting a service timeout.
The timeout-vote.yaml is not working and is replaced by
istio-gateway-vote-timeout.yaml. This appropriately applies the timeout
rule in the main routes of the VirtualService resource. The README is
modified to reflect the changes.

Removed explicit service namespace location on fault-injection.yaml.
Minor change so that yaml file would still work if user has a different
default namespace.

This commit also adds a label `app` to deploy-cloudant.yaml's pod
specification so that Istio's service graph will label them accordingly.

Signed-off-by: AnthonyAmanse <ghieamanse@gmail.com>